Move constants.py into RimSort.utils
