+++ RimSort/controllers/app_controller.py
@@ -7,7 +7,7 @@ from PySide6.QtCore import QObject
 from PySide6.QtWidgets import QApplication
 
 from RimSort.utils.app_info import AppInfo
-from util.constants import DEFAULT_USER_RULES
+from RimSort.utils.constants import DEFAULT_USER_RULES
 from RimSort.utils.system_info import SystemInfo
 from RimSort.views.main_window import MainWindow
