+++ RimSort/windows/missing_mods_panel.py
@@ -16,7 +16,7 @@ from PySide6.QtWidgets import (
     QVBoxLayout,
 )
 
-from util.constants import RIMWORLD_DLC_METADATA
+from RimSort.utils.constants import RIMWORLD_DLC_METADATA
 
 
 class MissingModsPrompt(QWidget):
